BBDD.sql : He quitado el rol "Invitado" a insertar, dejando a "Usuario" y "Administrador"
funcionesBBDD.php :
Le he dado un mejor formato al código y he añadido los comentarios pertinentes
Modificada función eliminarProyecto() para que devuelva un booleano indicando el resultado de su ejecución
Añadida función eliminarTarea(), que realizará las acciones pertinentes a la hora de que se decida eliminar una tarea
Añadida función eliminarSubtareas(), que realizará las acciones pertinentes a la hora que se decida eliminar todas las subtareas de una tarea

Utils.php :
Se ha movido el lugar de declaración de las constantes aquí

Se ha actualizado la info en el Readme

+++ Proyecto/Fuentes/php/funcionesBBDD.php
@@ -1,18 +1,11 @@
 <?php // En este archivo almacenaré las funciones que usaré para gestionar la base de datos
-    require_once "Utils.php"; // Linkeo el archivo de Utils a este, para usar sus funciones
 
-    const NOMBRE_BBDD = "todomanager"; // Constante indicando el nombre de la BBDD, para evitar números y texto mágico
-    // Constantes referentes a los nombres de las tablas de la BBDD
-    const TABLA_USUARIOS = "usuarios";
-    const TABLA_PRIVILEGIOS = "privilegios";
-    const TABLA_ROLES = "roles";
-    const TABLA_PROYECTOS = "proyectos";
-    const TABLA_TAREAS = "tareas";
-    const TABLA_TAREAS_FINALIZADAS = "tareas_finalizadas";
-    
+    require_once "Utils.php"; // Linkeo el archivo de Utils a este, para usar sus funciones
+  
 
     /**
-     * Intenta conectarse a la base de datos especificada y guarda en una variable el tipo de error en cada caso.
+     * Intenta conectarse a la base de datos especificada y guarda en una variable el tipo de error en cada caso
+     * 
      * @return La conexión a la BBDD si no hay errores, o false si ocurre algún error
      */
     function conectarBBDD(){
@@ -29,7 +22,10 @@
     }
 
 
-    // Funciones referentes a los usuarios :
+
+
+    // ----------------------------------------------------------- Funciones referentes a los usuarios ---------------------------------------------------------
+
     /**
      * Inserta los datos de un nuevo Usuario en la BBDD
      * 
@@ -39,9 +35,12 @@
      * @param $password Contraseña del nuevo usuario
      * @param $imagen base64 de la imagen del nuevo usuario
      * @param $rol Rol que va a tener el nuevo usuario
+     * 
+     * @return Boolean indicando si la acción resultó con errores
      */
     function registrarUsuario($conexion, $email, $nickname, $password, $imagen, $rol){ // TODO : El rol seguramente vaya a mano
-        $sentencia = "INSERT INTO ".TABLA_USUARIOS." (email, nickname, pwd, imagen, rol) VALUES(".$email.", ".$nickname.", ".$password.", ".$imagen.", ".$rol.")"; // Armo la sentencia
+        // Armo la sentencia
+        $sentencia = "INSERT INTO ".TABLA_USUARIOS." (email, nickname, pwd, imagen, rol) VALUES(".$email.", ".$nickname.", ".$password.", ".$imagen.", ".$rol.")";
         if (!mysqli_query($conexion, $sentencia)) { // Compruebo que al ejecutar la consulta haya salido algo mal
             consoleLog("Se ha producido un error al registrar al usuario : ".$conexion-> connect_error); // Y en ese caso muestro un mensaje de error en la consola
             $conexion-> close(); // Finalmente cierro la conexión a la base de datos
@@ -60,7 +59,7 @@
      * @param $conexion La conexión a la base de datos
      * @param $id La ID del usuario a eliminar
      * 
-     * @return Una booleana indicando si se ha realizado satisfactoriamente
+     * @return Boolean indicando si la acción resultó con errores
      */
     function eliminarUsuario($conexion, $id){
         $sentencia = "DELETE FROM ".TABLA_USUARIOS." WHERE id = ".$id; // Armo la sentencia
@@ -79,8 +78,11 @@
 
     /**
      * Elimina de la base de datos todos los proyectos y las tareas a cargo de un usuario
+     * 
      * @param $conexion La conexión con la base de datos
      * @param $idUsuario La ID del usuario sobre el que vamos a eliminar los proyectos
+     * 
+     * @return Boolean indicando si la acción resultó con errores
      */
     function eliminarProyectosDeUsuario($conexion, $idUsuario){
         $sentencia = "SELECT * FROM ".TABLA_PROYECTOS." WHERE usuario_creador = ".$idUsuario; // Consigo todos los proyectos de la base de datos
@@ -95,26 +97,47 @@
     }
 
 
+
+    // ----------------------------------------------------------- Funciones generales ---------------------------------------------------------
+
     /**
-     * Elimina de la base de datos todas las tareas de un proyecto
+     * Elimina un proyecto y todas sus tareas de la base de datos
+     * 
      * @param $conexion La conexión con la base de datos
-     * @param $proyecto El proyecto sobre el que se van a eliminar todas las tareas
+     * @param $proyecto El objeto del proyecto que va a ser eliminado, y sobre el que se van a eliminar todas las tareas
+     * 
+     * @return Boolean indicando si la acción resultó con errores
      */
     function eliminarProyecto($conexion, $proyecto){
-        eliminarTodasTareas($conexion, $proyecto);
-
-        // Si todo sale bien, armo la consulta para eliminar el proyecto
-        $sentencia = "DELETE FROM ".TABLA_PROYECTOS." WHERE id =".$proyecto-> id;
-        if (!mysqli_query($onexion, $sentencia)) {
-            // En caso incorrecto, muestro un mensaje de error en la consola
-            consoleLog("Se ha producido un error al intentar eliminar el proyecto ".$proyecto-> nombre." : ".$conexion-> connect_error);
-            $conexion-> close(); // Y cierro la conexión a la base de datos                
-        }        
+        if (eliminarTodasTareas($conexion, $proyecto)) { // Compruebo que haya salido bien la realización de esta función
+            // Si sale bien, armo la consulta para eliminar el proyecto
+            $sentencia = "DELETE FROM ".TABLA_PROYECTOS." WHERE id =".$proyecto-> id;
+            if (!mysqli_query($onexion, $sentencia)) {
+                // En caso incorrecto, muestro un mensaje de error en la consola
+                consoleLog("Se ha producido un error al intentar eliminar el proyecto ".$proyecto-> nombre." : ".$conexion-> connect_error);
+                $conexion-> close(); // Y cierro la conexión a la base de datos 
+                return false; // Devuelvo false indicando que algo ha ido mal    
+            }
+            else {
+                return true; // Devuelvo true si todo ha ido bien
+            }                    
+        }
+        else { // En el caso de que haya algún error
+            // Muestro un mensaje en la consola
+            consoleLog("Se ha producido un error al intentar eliminar todas las tareas del proyecto ".$proyecto-> nombre." : ".$conexion-> connect_error);
+            $conexion-> close(); // Y cierro la conexión a la base de datos  
+            return false; // Devuelvo false indicando que algo ha ido mal 
+        }
     }
 
 
     /**
+     * Elimina todas las tareas y tareas finalizadas de un proyecto
      * 
+     * @param $conexion La conexión con la base de datos
+     * @param $proyecto El objeto del proyecto sobre el que se van a eliminar todas sus tareas
+     * 
+     * @return Boolean indicando si la acción resultó con errores
      */
     function eliminarTodasTareas($conexion, $proyecto){
         // Armo la consulta para eliminar primero las tareas finalizadas del proyecto
@@ -138,8 +161,79 @@
             // En caso incorrecto, muestro un mensaje de error en la consola
             consoleLog("Se ha producido un error al intentar eliminar las tareas finalizadas correspondientes al proyecto ".$proyecto-> nombre." : ".$conexion-> connect_error);
             $conexion-> close(); // Y cierro la conexión a la base de datos
-
+            
             return false;
         }           
+    }
+
+
+    /**
+     * Elimina una tarea y todas sus subtareas
+     * 
+     * @param $conexion La conexión con la base de datos
+     * @param $tarea El objeto de la tarea que queremos eliminar
+     * 
+     * @return Boolean indicando si la acción resultó con errores
+     */
+    function eliminarTarea($conexion, $tarea){
+        // Primero elimino las subtareas de la tarea que se quiere eliminar
+        if (eliminarSubtareas($conexion, $tarea)) { 
+            // Armo la consulta       
+            $sentencia = "DELETE FROM ".TABLA_TAREAS." WHERE id=".$tarea-> id;
+
+            if (mysqli_query($conexion, $sentencia)) { // Intento eliminar la tarea
+                return true;
+            }
+            else {
+                // En caso incorrecto, muestro un mensaje de error en la consola
+                consoleLog("Se ha producido un error al intentar eliminar la tarea ".$tarea-> nombre." : ".$conexion-> connect_error);
+                $conexion-> close(); // Y cierro la conexión a la base de datos
+                    
+                return false;
+            }
+        }
+        else {
+            // En caso incorrecto, muestro un mensaje de error en la consola
+            consoleLog("Se ha producido un error al intentar eliminar las subtareas de la tarea ".$tarea-> nombre." : ".$conexion-> connect_error);
+            $conexion-> close(); // Y cierro la conexión a la base de datos
+                
+            return false;
+        }
+    }
+
+
+    /**
+     * Elimina todas las subtareas de una tarea
+     * 
+     * @param $conexion La conexión con la base de datos
+     * @param $tarea La tarea sobre la que queremos eliminar sus subtareas
+     * 
+     * @return Boolean indicando si la acción resultó con errores
+     */
+    function eliminarSubtareas($conexion, $tarea){
+        // Primero intento eliminar las subtareas finalizadas
+        $sentencia = "DELETE FROM ".TABLA_TAREAS_FINALIZADAS." WHERE parentID=".$tarea-> id;
+        if (mysqli_query($conexion, $sentencia)) { // Intento eliminar las subtareas finalizadas
+
+            // Ahora intento eliminar las subtareas activas
+            $sentencia = "DELETE FROM ".TABLA_TAREAS." WHERE parentID=".$tarea-> id;
+            if (mysqli_query($conexion, $sentencia)) { // Intento eliminar las subtareas activas
+                return true;
+            }
+            else {
+                // En caso incorrecto, muestro un mensaje de error en la consola
+                consoleLog("Se ha producido un error al intentar eliminar las subtareas de la tarea ".$tarea-> nombre." : ".$conexion-> connect_error);
+                $conexion-> close(); // Y cierro la conexión a la base de datos
+                
+                return false;                
+            }
+        }
+        else {
+            // En caso incorrecto, muestro un mensaje de error en la consola
+            consoleLog("Se ha producido un error al intentar eliminar las subtareas finalizadas de la tarea ".$tarea-> nombre." : ".$conexion-> connect_error);
+            $conexion-> close(); // Y cierro la conexión a la base de datos
+                
+            return false;
+        }
     }
 ?>
